+++ idEngine/UI/idListWindow.cs
@@ -104,7 +104,7 @@ namespace idTech4.UI
 			}*/
 		}
 
-		public override void Draw(int x, int y)
+		public override void Draw(float x, float y)
 		{
 			//idConsole.Warning("TODO: ListWindow Draw");
 			/*idVec4 color;
