+++ idEngine/UI/idRenderWindow.cs
@@ -96,7 +96,7 @@ namespace idTech4.UI
 
 		#region idWindow implementation
 		#region Public
-		public override void Draw(int x, int y)
+		public override void Draw(float x, float y)
 		{
 			//idConsole.Warning("TODO: RenderWindow Draw");
